pyproject.toml
[tool.poetry]
name = "invoive-generator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiosqlite = "0.21.0"
annotated-types = "0.7.0"
anyio = "4.8.0"
asyncpg = "0.30.0"
bcrypt = "4.2.1"
certifi = "2025.1.31"
click = "8.1.8"
distro = "1.9.0"
ecdsa = "0.19.0"
fastapi = "0.115.8"
filelock = "3.17.0"
fsspec = "2025.2.0"
h11 = "0.14.0"
httpcore = "1.0.7"
httpx = "0.28.1"
idna = "3.10"
jinja2 = "3.1.5"
jiter = "0.8.2"
markupsafe = "3.0.2"
mpmath = "1.3.0"
networkx = "3.4.2"
openai = "1.61.1"
passlib = "1.7.4"
pdf2image = "1.17.0"
pillow = "11.1.0"
psycopg2-binary = "2.9.10"
pyasn1 = "0.6.1"
pydantic = "2.10.6"
pydantic-core = "2.27.2"
pypdf2 = "3.0.1"
python-dateutil = "2.9.0.post0"
python-jose = "3.3.0"
python-multipart = "0.0.20"
rsa = "4.9"
scipy = "1.15.1"
setuptools = "75.8.0"
six = "1.17.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.38"
starlette = "0.45.3"
sympy = "1.13.1"
torch = "2.6.0"
torchvision = "0.21.0"
tqdm = "4.67.1"
typing-extensions = "4.12.2"
uvicorn = "0.34.0"
groq = "^0.18.0"
paddleocr = "^2.9.1"
python-dotenv = "^1.0.1"
aisqlite = "^1.1.0"
paddlepaddle = "^2.6.2"
pymysql = "^1.1.1"
aiomysql = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
